refactor(executor): extract fill_tracker.py — OrderAckInfo, FillSummary, fill polling
Commit 5/6: Keystone extraction — fill tracking subsystem.

Create execution/fill_tracker.py containing:
- OrderAckInfo, FillSummary dataclasses
- fetch_order_status, fetch_order_trades (API helpers)
- emit_order_ack, should_emit_close, confirm_order_fill

executor_live.py: -355 LOC (6053 → 5698)
fill_tracker.py: +366 LOC (new file)

Coupling change: _POSITION_TRACKER is now instantiated in
fill_tracker.py and re-exported to executor_live.py. The
confirm_order_fill function accepts an optional position_tracker
parameter for future DI use.

Also removes 4 duplicate helper re-definitions (normalize_status,
to_float, ms_to_iso, iso_to_ts) that were local copies of the
already-imported execution.helpers functions.

Removes unused dataclasses import from executor_live.py.

Test patch paths updated: test_executor_events.py now patches
fill_tracker module directly.

Test: 3949 passed, 46 skipped
